requirements.txt
pytest==8.3.3
rich==13.8.1

# not used currently
pydantic>2.7.1
pyyaml==6.0.2
dict2xml==1.7.6